pyproject.toml
[project]
name = "data-agent-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.57.1",
    "duckdb>=1.5.2",
    "levenshtein>=0.27.3",
    "openai>=2.31.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.13.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]

[dependency-groups]
dev = [
    "hypothesis>=6.151.13",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]
